Added settings for Boosts/MouseHighlight Changed overlay colors to fit "RS" style
