--- runelite-client/src/main/java/net/runelite/client/util/Utilities.java
@@ -1,68 +1,0 @@
-/*
- * Copyright (c) 2017, Adam <Adam@sigterm.info>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package net.runelite.client.util;
-
-import java.awt.*;
-
-
-public class Utilities
-{
-
-	public static void fillRect(Graphics g, int x, int y, int w, int h)
-	{
-		Color color = g.getColor();
-		//Rect
-		g.setColor(color);
-		g.fillRect(x, y, w, h);
-		//Outline
-		g.setColor(color.darker());
-		g.drawRect(x, y, w, h);
-	}
-
-	public static void outlineString(Graphics g, String text, int x, int y)
-	{
-		Color color = g.getColor();
-		//Outline
-		g.setColor(Color.BLACK);
-		g.drawString(text, x + 1, y + 1);
-		g.drawString(text, x - 1, y - 1);
-		g.drawString(text, x + 1, y - 1);
-		g.drawString(text, x - 1, y + 1);
-		//String
-		g.setColor(color);
-		g.drawString(text, x, y);
-	}
-	public static void shadowString(Graphics g, String text, int x, int y)
-	{
-		Color color = g.getColor();
-		//Shadow
-		g.setColor(Color.BLACK);
-		g.drawString(text, x + 1, y + 1);
-		//String
-		g.setColor(color);
-		g.drawString(text, x, y);
-	}
-
-}
